move fsbpfinding query to common
