+++ packages/cloudbuster/src/findings.ts
@@ -1,33 +1,11 @@
-import type { aws_securityhub_findings, PrismaClient } from '@prisma/client';
-import type { Finding, GroupedFindings, SecurityHubSeverity } from './types';
-
-/**
- * Queries the database for FSBP findings
- */
-export async function getFsbpFindings(
-	prisma: PrismaClient,
-	severities: SecurityHubSeverity[],
-): Promise<Finding[]> {
-	const findings = await prisma.aws_securityhub_findings.findMany({
-		where: {
-			OR: severities.map((s) => ({
-				severity: { path: ['Label'], equals: s },
-			})),
-			AND: {
-				generator_id: {
-					startsWith: 'aws-foundational-security-best-practices/v/1.0.0',
-				},
-			},
-		},
-	});
-
-	return findings.map(transformFinding);
-}
+import type { aws_securityhub_findings } from '@prisma/client';
+import type { SecurityHubSeverity } from 'common/src/types';
+import type { Finding, GroupedFindings } from './types';
 
 /**
  * Transforms a SQL row into a finding
  */
-function transformFinding(finding: aws_securityhub_findings): Finding {
+export function transformFinding(finding: aws_securityhub_findings): Finding {
 	let severity = null;
 	let priority = null;
 	let remediationUrl = null;
