+++ packages/refresh-materialized-view/src/config.ts
@@ -1,10 +1,10 @@
-import type { DatabaseConfig, PrismaConfig } from 'common/database';
+import { getEnvOrThrow } from 'common/functions';
+import type { DatabaseConfig, PrismaConfig } from 'common/src/database-setup';
 import {
 	getDatabaseConfig,
 	getDatabaseConnectionString,
 	getDevDatabaseConfig,
-} from 'common/database';
-import { getEnvOrThrow } from 'common/functions';
+} from 'common/src/database-setup';
 
 export interface Config extends PrismaConfig {
 	/**
