+++ packages/repocop/src/config.ts
@@ -1,11 +1,11 @@
 import * as process from 'process';
+import { getEnvOrThrow } from 'common/functions';
 import {
 	getDatabaseConfig,
 	getDatabaseConnectionString,
 	getDevDatabaseConfig,
-} from 'common/database';
-import type { DatabaseConfig, PrismaConfig } from 'common/database';
-import { getEnvOrThrow } from 'common/functions';
+} from 'common/src/database-setup';
+import type { DatabaseConfig, PrismaConfig } from 'common/src/database-setup';
 
 export interface Config extends PrismaConfig {
 	/**
